backend-quikscribe/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic>=1.16.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
authlib==1.2.1
python-multipart==0.0.6

# Validation & Serialization
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Configuration
python-dotenv==1.0.0

# Session Management
itsdangerous==2.1.2

# Docker Integration
docker==6.1.3

# Kubernetes client (for creating Jobs from backend)
kubernetes==26.1.0